refactor(docs-server): split data_fetcher.py into focused modules and add fallback update script
Address PR reviewer feedback by extracting data_fetcher.py (823 lines) into
four focused modules: http.py (HTTP infrastructure and caching), sitemap.py
(sitemap discovery), html_processing.py (HTML-to-markdown conversion), and a
slimmed data_fetcher.py (business logic). Also adds scripts/update_fallback_constants.py
to regenerate hardcoded fallback values from the live sitemap.
